@@ -9,22 +9,21 @@
   <packaging>pom</packaging>
 
   <properties>
+    <compiler-plugin.version>3.11.0</compiler-plugin.version>
     <maven.compiler.parameters>true</maven.compiler.parameters>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
-    <maven.compiler.forceJavacCompilerUse>true</maven.compiler.forceJavacCompilerUse>
+    <maven.compiler.source>17</maven.compiler.source>
+    <maven.compiler.target>17</maven.compiler.target>
+
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
 
-    <quarkus-plugin.version>2.6.3.Final</quarkus-plugin.version>
+    <quarkus-plugin.version>3.1.0.Final</quarkus-plugin.version>
     <quarkus.platform.artifact-id>quarkus-bom</quarkus.platform.artifact-id>
     <quarkus.platform.group-id>io.quarkus.platform</quarkus.platform.group-id>
-    <quarkus.platform.version>2.6.2.Final</quarkus.platform.version>
+    <quarkus.platform.version>3.1.0.Final</quarkus.platform.version>
 
-    <testcontainers.version>1.16.3</testcontainers.version>
-
-    <compiler-plugin.version>3.9.0</compiler-plugin.version>
-    <surefire-plugin.version>2.22.2</surefire-plugin.version>
+    <surefire-plugin.version>3.0.0</surefire-plugin.version>
+    <skipITs>true</skipITs>
   </properties>
 
   <dependencyManagement>
@@ -42,5 +41,6 @@
   <modules>
     <module>quarkus-todo</module>
     <module>quarkus-todo-reactive</module>
+    <module>quarkus-todo-loom</module>
   </modules>
 </project>
