@@ -4,9 +4,9 @@ import io.quarkus.hibernate.reactive.panache.PanacheEntity;
 import io.smallrye.mutiny.Multi;
 import io.smallrye.mutiny.Uni;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.validation.constraints.NotBlank;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.validation.constraints.NotBlank;
 import java.util.List;
 
 @Entity
@@ -23,12 +23,12 @@ public class Todo extends PanacheEntity {
 
     public String url;
 
-    public static Multi<Todo> findNotCompleted() {
-        return stream("completed", false);
+    public static Uni<List<Todo>> findNotCompleted() {
+        return list("completed", false);
     }
 
-    public static Multi<Todo> findCompleted() {
-        return stream("completed", true);
+    public static Uni<List<Todo>> findCompleted() {
+        return list("completed", true);
     }
 
     public static Uni<Long> deleteCompleted() {
