@@ -1,29 +1,18 @@
 package io.quarkus.sample;
 
-import java.util.List;
-import javax.validation.Valid;
-import javax.ws.rs.Consumes;
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.OPTIONS;
-import javax.ws.rs.PATCH;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-
 import io.quarkus.hibernate.reactive.panache.Panache;
+import io.quarkus.hibernate.reactive.panache.PanacheEntityBase;
 import io.quarkus.panache.common.Sort;
-import io.smallrye.common.annotation.Blocking;
 import io.smallrye.mutiny.Uni;
+import jakarta.validation.Valid;
+import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response.Status;
+
+import java.util.List;
 
 @Path("/api")
-@Produces(MediaType.APPLICATION_JSON)
-@Consumes(MediaType.APPLICATION_JSON)
 public class TodoResource {
 
     @OPTIONS
@@ -33,17 +22,14 @@ public class TodoResource {
 
     @GET
     public Uni<List<Todo>> getAll() {
-        System.out.println(Thread.currentThread().getName());
         return Panache.withTransaction(
                 () -> Todo.findAll(Sort.by("order")).list()
         );
     }
 
     @GET
-    @Blocking
     @Path("/blocking")
     public List<Todo> getAllBlocking() {
-        System.out.println(Thread.currentThread().getName());
         return getAll()
                 .await().indefinitely();
     }
@@ -53,18 +39,17 @@ public class TodoResource {
     public Uni<Todo> getOne(@PathParam("id") Long id) {
         return Panache.withTransaction(
                 () -> Todo.<Todo>findById(id)
-                    .onItem().ifNull().failWith(() ->
-                        new WebApplicationException("Todo with id of " + id + " does not exist.", Status.NOT_FOUND)
-                ));
+                        .onItem().ifNull().failWith(() ->
+                                new WebApplicationException("Todo with id of " + id + " does not exist.", Status.NOT_FOUND)
+                        ));
     }
 
     @POST
     public Uni<Response> create(@Valid Todo item) {
-        return Panache.withTransaction(
-                () -> item.persist()
-            ).replaceWith(
-                () -> Response.status(Status.CREATED).entity(item).build()
-        );
+        return Panache.withTransaction(item::persist)
+                .replaceWith(
+                        () -> Response.status(Status.CREATED).entity(item).build()
+                );
     }
 
     @PATCH
@@ -72,21 +57,21 @@ public class TodoResource {
     public Uni<Todo> update(@Valid Todo todo, @PathParam("id") Long id) {
         return Panache.withTransaction(
                 () -> Todo.<Todo>findById(id)
-                .onItem().transform(entity -> {
-                    entity.id = id;
-                    entity.completed = todo.completed;
-                    entity.order = todo.order;
-                    entity.title = todo.title;
-                    entity.url = todo.url;
-                    return entity;
-                })
+                        .onItem().transform(entity -> {
+                            entity.id = id;
+                            entity.completed = todo.completed;
+                            entity.order = todo.order;
+                            entity.title = todo.title;
+                            entity.url = todo.url;
+                            return entity;
+                        })
         );
     }
 
     @DELETE
     public Uni<Response> deleteCompleted() {
         return Panache.withTransaction(
-                () -> Todo.deleteCompleted()
+                Todo::deleteCompleted
         ).replaceWith(
                 () -> Response.noContent().build()
         );
@@ -98,9 +83,9 @@ public class TodoResource {
         return Panache.withTransaction(
                 () -> Todo.findById(id)
                         .onItem().ifNull().failWith(() ->
-                            new WebApplicationException("Todo with id of " + id + " does not exist.", Status.NOT_FOUND)
+                                new WebApplicationException("Todo with id of " + id + " does not exist.", Status.NOT_FOUND)
                         )
-                        .call(entity -> entity.delete())
+                        .call(PanacheEntityBase::delete)
         ).replaceWith(
                 () -> Response.noContent().build()
         );
